Today I learnt the difference between "=" and "==".
"=" means we initialize a variable or assign a new value to an existing variable.
"==" means the comparision between the variable we initialized and something else.
"!=" is the opposite operation of "==".

One more thing, if we type "int c;" without initialized value, c could be anything! It's very intersting how the computer actually generate a random number, though I still need more practice and spend more time to completely understand the mechniasm underneath.