Today I learnt if and else
I tried to implement a function, which is to get the sum of absolute value of 3 integers. Also I learnt a weird assignment a = -a; they both are variables, but it can be changed in low-level binary? Still not 100% get it...

Why function call need a ';' but if does not? Though I know the answer is 'if' is C built-in syntax and function call is not.