Example 20: A monkey picked a certain number of peaches on the first day. It immediately ate half of them and, still not satisfied, ate one more. The next morning, it ate half of the remaining peaches and again ate one more. Every morning after that, it ate half of the peaches left from the previous day plus one more. When it wanted to eat on the morning of the tenth day, there was only one peach left. Find out how many peaches were picked on the first day.
Program Analysis: Adopt the method of reverse thinking to infer from the back to the front.

So I basically first created a varable named x and put it into a for loop and made the for loop go ten times with adding 1 and then timing 2 to x => lastly printf x.
Another solution I tried will print out every peaches left peer a day - this one is basicallt the same as the first one just that I used a 1D array.